+++ tests-mac/FontRenderingTest/AppDelegate.m
@@ -115,7 +115,7 @@
     [button addTarget:self action:@selector(buttonLeftMouseDown:) forControlEvents:ICControlEventLeftMouseDown];
     [button addTarget:self action:@selector(buttonLeftMouseUpInside:) forControlEvents:ICControlEventLeftMouseUpInside];
     
-    ICTextField *textField = [[[ICTextField alloc] initWithSize:icSizeMake(200, 200)] autorelease];
+    ICTextField *textField = [[[ICTextField alloc] initWithSize:icSizeMake(400, 500)] autorelease];
     [textField setFont:[ICFont fontWithName:@"Arial" size:13]];
     [textField setPositionY:300];
     [textField setText:@"Text field\nTest\nwith multiple lines"];